refactor: guard 및 strategy 파일이 동
