+++ src/app.module.ts
@@ -6,7 +6,7 @@ import { DataSource } from 'typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
-import { JwtAuthGuard } from './auth/jwt-auth.guard';
+import { JwtAuthGuard } from './auth/guards/jwt-auth.guard';
 import { RedisModule } from './redis.module';
 import { UsersModule } from './users/users.module';
 @Module({
